# src/main/resources/openapi/userupdate.yaml
openapi: 3.0.0
info:
  title: Title
  description: Title
  version: 1.0.0
servers:
  - url: 'http'
paths:
  /users/{userId}:
    put:
      summary: Update a user
      operationId: updateUser
      parameters:
        - name : userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: User details that need to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequest"
      responses:
        '200':
          description: "ok"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        '400':
          description: "invalid id"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '401':
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '403':
          description: "forbidden"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '404':
          description: "not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    UserResponse:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
        email:
          type: string
    UserRequest:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
        email:
          type: string
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
  